dom/nodes/moveBefore/continue-css-animation.tentative.html

<!DOCTYPE html>
<title>Node.moveBefore should preserve CSS animation state</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>
  <section id="old-parent">
    <div id="item"></div>
  </section>
  <section id="new-parent">
  </section>
  <style>
    @keyframes anim {
      from {
        transform: translateX(100px);
      }

      to {
        transform: translateX(400px);
      }
    }

    #item {
      width: 100px;
      height: 100px;
      background: green;
      animation: 1s linear infinite alternate anim;
      animation-delay: 100ms;
    }
  </style>
  <script>
    promise_test(async t => {
      const item = document.querySelector("#item");
      // Wait for the animation to start
      while (getComputedStyle(item).transform === "none") {
        await new Promise(resolve => requestAnimationFrame(() => resolve()));
      }

      // Reparent item
      document.body.querySelector("#new-parent").moveBefore(item, null);

      await new Promise(resolve => requestAnimationFrame(() => resolve()));

      assert_not_equals(getComputedStyle(item).transform, "none");
    });
  </script>
</body>
